+++ src/transactions/dex/DeleteOfferOpFrame.cpp
@@ -3,12 +3,13 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "DeleteOfferOpFrame.h"
+#include "OfferManager.h"
 #include "database/Database.h"
-#include "ledger/LedgerDelta.h"
+#include "ledger/BalanceHelperLegacy.h"
 #include "ledger/AssetPairHelper.h"
+#include "ledger/LedgerDelta.h"
 #include "ledger/OfferHelper.h"
 #include "main/Application.h"
-#include "OfferManager.h"
 
 namespace stellar
 {
@@ -27,7 +28,9 @@ DeleteOfferOpFrame::doApply(Application& app, LedgerDelta& delta,
                             LedgerManager& ledgerManager)
 {
     auto& db = app.getDatabase();
-    const auto offer = OfferHelper::Instance()->loadOffer(getSourceID(), mManageOffer.offerID, mManageOffer.orderBookID, db, &delta);
+    const auto offer = OfferHelper::Instance()->loadOffer(
+        getSourceID(), mManageOffer.offerID, mManageOffer.orderBookID, db,
+        &delta);
     if (!offer)
     {
         innerResult().code(ManageOfferResultCode::NOT_FOUND);
@@ -35,14 +38,22 @@ DeleteOfferOpFrame::doApply(Application& app, LedgerDelta& delta,
     }
 
     OfferManager::deleteOffer(offer, db, delta);
+    auto balanceHelper = BalanceHelperLegacy::Instance();
+    auto baseBalance =
+        balanceHelper->mustLoadBalance(offer->getOffer().baseBalance, db, &delta);
+    auto quoteBalance = balanceHelper->mustLoadBalance(
+        offer->getOffer().quoteBalance, db, &delta);
     innerResult().code(ManageOfferResultCode::SUCCESS);
-    innerResult().success().offer.effect(ManageOfferEffect::DELETED);
-
+    auto& success = innerResult().success();
+    success.offer.effect(ManageOfferEffect::DELETED);
+    success.baseAsset = baseBalance->getAsset();
+    success.quoteAsset = quoteBalance->getAsset();
     return true;
 }
 
 // makes sure the currencies are different
-bool DeleteOfferOpFrame::doCheckValid(Application& app)
+bool
+DeleteOfferOpFrame::doCheckValid(Application& app)
 {
     if (mManageOffer.amount != 0)
     {
@@ -54,7 +65,7 @@ bool DeleteOfferOpFrame::doCheckValid(Application& app)
         innerResult().code(ManageOfferResultCode::NOT_FOUND);
         return false;
     }
-    
+
     return true;
 }
-}
+} // namespace stellar